+++ .idea/.name
@@ -1,1 +1,1 @@
-GeoGoal
+GeoGoal-master
