+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/GeoGoal.iml" filepath="$PROJECT_DIR$/GeoGoal.iml" />
+      <module fileurl="file://$PROJECT_DIR$/GeoGoal-master.iml" filepath="$PROJECT_DIR$/GeoGoal-master.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
