--- GeoGoal-master/app/src/main/java/bgb/geogoal/Background.java
@@ -1,34 +1,0 @@
-package bgb.geogoal;
-
-import android.graphics.Bitmap;
-import android.graphics.Canvas;
-
-public class Background {
-
-    private Bitmap image;
-    private int x, y, dx;
-
-    public Background(Bitmap res)
-    {
-        image = res;
-    }
-    public void update()
-    {
-        x+=dx;
-        if(x<0){
-            x=0;
-        }
-    }
-    public void draw(Canvas canvas)
-    {
-        canvas.drawBitmap(image, x, y,null);
-        if(x<0)
-        {
-            canvas.drawBitmap(image, x+GamePanel.WIDTH, y, null);
-        }
-    }
-    public void setVector(int dx)
-    {
-        this.dx = dx;
-    }
-}
