--- GeoGoal-master/app/src/main/java/bgb/geogoal/Game.java
@@ -1,26 +1,0 @@
-package bgb.geogoal;
-
-import android.app.Activity;
-import android.os.Bundle;
-import android.view.Window;
-import android.view.WindowManager;
-
-
-public class Game extends Activity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        //turn title off
-        requestWindowFeature(Window.FEATURE_NO_TITLE);
-
-        //set to full screen
-        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,WindowManager.LayoutParams.FLAG_FULLSCREEN);
-
-        setContentView(new GamePanel(this));
-    }
-
-
-
-}
