--- GeoGoal-master/app/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">GeoGoal</string>
-    <string name="title_activity_game">Game</string>
-</resources>
